refactor: move getDefaultErrorPage into ResponseBodyHandler
